Task 1) RAG and Text-to-SQL/requirements.txt
streamlit
python-dotenv
nest-asyncio
llama_index
sqlalchemy
pydantic
llama-cloud
openai
